+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## next
 
 - Add `widget::Indicator` to draw a single indicator
+- Rename `widget::RenderIndicator` to `widget::OrIndicators`
 
 ## 0.1.0